+++ src/components/scenes/MigrateWalletCompletionScene.tsx
@@ -1,4 +1,4 @@
-import { add, sub } from 'biggystring'
+import { add, lt, sub } from 'biggystring'
 import {
   EdgeCurrencyWallet,
   EdgeSpendInfo,
@@ -15,6 +15,7 @@ import {
   writeSyncedSettings
 } from '../../actions/SettingsActions'
 import { SCROLL_INDICATOR_INSET_FIX } from '../../constants/constantSettings'
+import { getSpecialCurrencyInfo } from '../../constants/WalletAndCurrencyConstants'
 import { useAsyncEffect } from '../../hooks/useAsyncEffect'
 import { useHandler } from '../../hooks/useHandler'
 import { useWatch } from '../../hooks/useWatch'
@@ -22,11 +23,12 @@ import { lstrings } from '../../locales/strings'
 import { useSelector } from '../../types/reactRedux'
 import { EdgeAppSceneProps } from '../../types/routerTypes'
 import { getWalletName } from '../../util/CurrencyWalletHelpers'
+import { convertNativeToDenomination } from '../../util/utils'
+import { SceneButtons } from '../buttons/SceneButtons'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
+import { useTheme } from '../services/ThemeContext'
 import { CreateWalletSelectCryptoRow } from '../themed/CreateWalletSelectCryptoRow'
-import { MainButton } from '../themed/MainButton'
 import { SceneHeader } from '../themed/SceneHeader'
 import { MigrateWalletItem } from './MigrateWalletSelectCryptoScene'
 
@@ -45,7 +47,6 @@ const MigrateWalletCompletionComponent = (props: Props) => {
   const { migrateWalletList } = route.params
 
   const theme = useTheme()
-  const styles = getStyles(theme)
 
   const account = useSelector(state => state.core.account)
   const defaultIsoFiat = useSelector(state => state.ui.settings.defaultIsoFiat)
@@ -244,6 +245,35 @@ const MigrateWalletCompletionComponent = (props: Props) => {
                 ]
               }
               const amountToSend = sub(maxAmount, feeTotal)
+
+              // Validate that the new wallet will have sufficient balance after
+              // migration to meet minimum balance requirements
+              const { currencyInfo } = newWallet
+              const specialCurrencyInfo = getSpecialCurrencyInfo(
+                currencyInfo.pluginId
+              )
+              const minimumNativeBalance =
+                specialCurrencyInfo.minimumPopupModals?.minimumNativeBalance ??
+                '0'
+
+              if (lt(amountToSend, minimumNativeBalance)) {
+                const denom = currencyInfo.denominations.find(
+                  d => d.name === currencyInfo.currencyCode
+                )
+                if (denom != null) {
+                  const exchangeAmount = convertNativeToDenomination(
+                    denom.multiplier
+                  )(minimumNativeBalance)
+                  throw new Error(
+                    sprintf(
+                      lstrings.migrate_wallet_below_minimum_balance_2s,
+                      exchangeAmount,
+                      currencyInfo.currencyCode
+                    )
+                  )
+                }
+              }
+
               spendInfo = {
                 ...spendInfo,
                 spendTargets: [
@@ -330,26 +360,6 @@ const MigrateWalletCompletionComponent = (props: Props) => {
     }
   )
 
-  const renderNextButton = React.useMemo(() => {
-    return (
-      <View style={styles.bottomButton}>
-        <MainButton
-          spinner={!done}
-          disabled={!done}
-          label={!done ? undefined : lstrings.string_done_cap}
-          type="secondary"
-          marginRem={[0, 0, 1]}
-          onPress={() =>
-            navigation.navigate('edgeTabs', {
-              screen: 'walletsTab',
-              params: { screen: 'walletList' }
-            })
-          }
-        />
-      </View>
-    )
-  }, [done, navigation, styles.bottomButton])
-
   const keyExtractor = useHandler((item: MigrateWalletItem) => item.key)
 
   return (
@@ -374,20 +384,22 @@ const MigrateWalletCompletionComponent = (props: Props) => {
             scrollEnabled={done}
             scrollIndicatorInsets={SCROLL_INDICATOR_INSET_FIX}
           />
-          {renderNextButton}
+          <SceneButtons
+            primary={{
+              label: lstrings.string_done_cap,
+              disabled: !done,
+              onPress: () =>
+                navigation.navigate('edgeTabs', {
+                  screen: 'walletsTab',
+                  params: { screen: 'walletList' }
+                })
+            }}
+          />
         </View>
       )}
     </SceneWrapper>
   )
 }
-
-const getStyles = cacheStyles((theme: Theme) => ({
-  bottomButton: {
-    alignSelf: 'center',
-    bottom: theme.rem(1),
-    position: 'absolute'
-  }
-}))
 
 const makeSpendSignAndBroadcast = async (
   wallet: EdgeCurrencyWallet,