+++ CHANGELOG.md
@@ -2,10 +2,29 @@
 
 ## Unreleased (develop)
 
-## 4.31.0 (staging)
+- added: `ReturnKeyTypeButton` to `FlipInputModal2`
+
+## 4.32.0 (staging)
+
+- added: Support for Paybis ACH buy
+- added: Private View Key wallet menu option for Zano
+- added: Support for `SwapBelowLimitError` and `SwapAboveLimitError` without limit amount props
+- added: `ReturnKeyTypeButton` to `FlipInputModal2`
+- changed: Reinstate "Sell" in `MenuTabs` for UK
+- changed: `NotificationCard` visibility rules updated to be less invasive
+- changed: Unsupported FIO assets removed from `ConnectWallets` and `ShareButtons`
+- changed: `WalletListMenuModal` now supports displaying loading states for async options
+- fixed: Fixed viewport issues on ChnageMiningFeeScene.
+- fixed: Account for minimum account balances when migrating/sweeping
+- fixed: Button UI styling in `MigrateWalletCompletionScene`
+- fixed: Map from FIO codes to Edge-supported currencies more correctly
+- remove: Wire transfer fiat buy
+
+## 4.31.0 (2025-07-07)
 
 - added: "Change Username" setting
 - added: Support for Zano alias name resolution.
+- added: Add `walletAddress` param to Moonpay buy quotes.
 - changed: Don't allow multiple pending EVM transactions.
 - fixed: Specify build target in deploy.ts to fix Apple silicon build environments
 - fixed: Maestro tests to successfully run again